ceph: create volumeID mapping configmap
cephcsi needs new volumeID mapping configmap
to store the new volumeID mappings in the configmap.

Incase of the Async DR the cluster id and pool id
might not be the same, when workload is moved from
one cluster to another cluster cephcsi need to
regenerate the omap data and new volume handler.
the omap controller will updates the configmap
with volumeid mapping details. cephcsi will use
the new volumeID when it exists.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>